+++ Chapter01/Exercises/Exercise101/Program.cs
@@ -1,3 +1,12 @@
 ﻿using System;
 
-Console.WriteLine("Hello World!");
+namespace Exercise101
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
