+++ Chapter01/Exercises/Exercise112/Program.cs
@@ -2,34 +2,28 @@
 using System.IO;
 using System.Threading.Tasks;
 
-public class Program
+using (var file = new StreamWriter("products.csv", append: true))
 {
-    public static async Task Main()
+    file.Write("\nOne more macbook without details.");
+}
+
+using (var fileStream = new FileStream("products.csv", FileMode.Open,
+    FileAccess.Read))
+{
+    await ReadFile(fileStream);
+}
+
+static async Task ReadFile(FileStream fileStream)
+{
+    using (var reader = new StreamReader(fileStream))
     {
-        using (var file = new StreamWriter("products.csv", append: true))
+        var content = await reader.ReadToEndAsync();
+
+        var lines = content.Split(Environment.NewLine);
+
+        foreach (var line in lines)
         {
-            file.Write("\nOne more macbook without details.");
-        }
-
-        using (var fileStream = new FileStream("products.csv", FileMode.Open,
-            FileAccess.Read))
-        {
-            await ReadFile(fileStream);
-        }
-    }
-
-    private static async Task ReadFile(FileStream fileStream)
-    {
-        using (var reader = new StreamReader(fileStream))
-        {
-            var content = await reader.ReadToEndAsync();
-
-            var lines = content.Split(Environment.NewLine);
-
-            foreach (var line in lines)
-            {
-                Console.WriteLine(line);
-            }
+            Console.WriteLine(line);
         }
     }
 }
