+++ Chapter01/Exercises/Exercise112/products.csv
@@ -1,5 +1,5 @@
-One more macbook without details.
-een;Condition;Price USD
+Model;Memory;Storage;Ports;Screen;Condition;Price USD
 Macbook Pro Mid 2012;8GB;500GB HDD;USB 2.0x2;13" screen;Refurbished;400
 Macbook Pro Mid 2014;8GB;512GB SSD;USB 3.0x3;15" screen;Refurbished;750
 Macbook Pro Late 2019;16GB;512GB SSD;USB 3.0x3;15" screen;Refurbished;1250
+One more macbook without details.
